Implemented bilateral filter using opencv

+++ convergence_map_analysis/include/bilateral.hpp
@@ -1,6 +1,6 @@
 #ifndef __BILATERAL_HPP
 #define __BILATERAL_HPP
 
-void bilateral_smooth(float *,long,float,float);
+void bilateral_smooth(float *,long,double,double);
 
 #endif
